Don't give up when encountering lines in collision mesh data.
Instead just skipping them.

As this makes the isMeshPrimitiveValid() check obsolete, I removed the
whole implementation of it (which, after all, was just a very verbose
way to print the mesh primitive name, and could be done in a single
expression instead). Furthermore, a call to this function was also the
only reason why addStage() too the collision mesh group parameter, so I
removed it from there as well.

The check is now moved to inside constructBulletSceneFromMeshes(), where
it just skips unsupported primitives with a warning. If the code could
access the originating Magnum::Trade::MeshData instead of
CollisionMeshData, it could be easily extended to support also triangle
fans and triangle strips (since there's a MeshTools utility for
converting them to triangle list), but right now it skips those as well.